add jwt verifier to middleware
