+++ pyproject.toml
@@ -10,7 +10,7 @@ dependencies = [
     "beautifulsoup4>=4.12.3",
     "httpx>=0.28.0",
     "mcp>=1.8.0,<2.0.0",
-    "fastmcp>=2.3.4,<2.4.0",
+    "fastmcp>=2.12.3",
     "python-dotenv>=1.0.1",
     "markdownify>=0.11.6",
     "markdown>=3.7.0",
